+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Release Versions:
 - feat(build): handle installation and linking of dependencies for pinocchio collision support (#161)
 - build: add missing licenses (#170)
 - ci: use caching from docker to run tests in CI (#429)
+- feat: integrate collision detection feature into robot model (#163)
 
 ## 7.4.0